Fix an issue where the bubble overflow wouldn't animate when collapsing
When we collapse we'd update the selected state from the overflow to
the top bubble. This would update mExpandedView in stackView which
would mean the top bubble view would be told to animate instead of
the overflow even though it's still what's visually selected.

This changes the selected state to be set to the top bubble the
next time the stack is expanded instead of while its collapsing.

Bug: 196444758
Test: atest BubbleDataTest
Test: manual - navigate to the bubble overflow and tap outside, verify
               that the overflow animates out just like if you were
               collapsing from a convo bubble.
Change-Id: I1182d48916013dcc8aa15fb188543d8689b1f8a6