Remove Task#mForceNotOrganized
The flag was added to avoid flickering when switching from PiP to
fullscreen with display orientation change. The actual problem is
that there may have a latency for fullscreen task organizer to
configure the surface. The original approach works because it will
call WindowContainer#migrateToNewSurfaceControl when sending vanished
event to task organizer, which will create a new surface so the old
transformation is dropped. But that is unnecessary to use a error-prone
flag (e.g. not reset) to resolve the case. It is enough to clear the
matrix and crop of the task surface.

Bug: 196233750
Bug: 179235349
Test: Expand PiP from portrait display to fullscreen landscape
Change-Id: I675f93284e52ee4a84568265827c7e8c6066273d