Avoid skipping creation of emulated volume for cloned profile user
As of now, emulated volume is not getting created for the user if media
provider instance is not running in that user space. But, in case of
cloned profile user, where media provider instance is not running and
media queries get redirected to the owner user space, we need to
facilitate the creation of emulated volume. This CL takes care of this
usecase.

Bug: 197310110
Test: manual on Pixel5
Change-Id: I52cd40ed47df84a71c9c17cb2b51dc4a23d0d56d