Update javadoc of CarrierConfigManager
The implementation in CarrierConfigLoader allows system privileged
app to access the API CarrierConfigManager#notifyConfigChangedForSubId,
starting from N release. Several features has been built on it.
It is better to explicitly specify the behavior in javadoc to avoid
misunderstanding.

For the similar reason, the javadoc of methods below are also updated:
- getConfigForSubId
- getConfig
- getConfigByComponentForSubId

The cl also removes the closing </p> tag which is not allowed by
Android API Guidelines (go/android-api-guidelines).

Bug: 204712404
Test: m offline-sdk-docs
Change-Id: I7da035937a0472194a8c8cbc5ef99481351f6d36