Make GlobalActionsDialogLite directly injectable
When GlobalActionsDialogLite was made injectable by
commit 2ba937133223398138a1f0cc7fcb89c3d6642219, it was
in the form of a dagger.Lazy. This was probably a mistake, as
a local reference to the content of this dagger.Lazy was also
kept, and this messes up the lifecycle and may, in strange
cases, allow the use of a Dialog that has been destroyed,
with consequences that are difficult to predict.

A direct injection of the GlobalActionsDialogLite is
probably more appropriate.

Test: manual
Change-Id: I1646c883b9a15db737bfc897c41dd26fb584d539