Fix bug in insets animation when hw acceleration unavailable
Since 1955089 'Apply Insets animation when hw acceleration unavailable'
has been applied, apps running without hardware acceleration have been
able to animate without any frame sync for now.

However, unlike the intention, the animation may not be performed
forever. Because the frame number -1 was intended as a very early frame
number that guarantees to show. But native BBQ accepts it as the
maximum frame number which can not be reached permanently for showing.
Because native BBQ takes the value as unsigned long long type (signed
-1 to unsigned maximum).

This change involves the following:
- Fix the misuse of the very early framenumber in
  ViewRootInsetsControllerHost.
- Treats all -ve values as 0 so BBQ doesn't cast it incorrectly.

Bug: 206482421
Signed-off-by: Hyeongseop Shim <hyeongseop.shim@samsung.corp-partner.google.com>
Change-Id: I87faf440fc657c43fd67caa8644ecbc971187d84